pyproject.toml
[tool.poetry]
name = "geodata-backend"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
black = "^23.12.1"
pylint = "^3.0.3"
psycopg2 = "^2.9.9"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
